Working on making progam cards work with movement
Not done yet. I've added a rotate enum and implemented it with program cards as well as player.
A player can now hold cards and I'm working on implementing the process of choosing which cards to use for the next round.
Added cards to gameScreen (all the cards in the game are there).