Update program register GUI
Added new ProgramRegisters texture
Changed life token from a circle to a heart to better reflect what kind of token it is
Added 'wires' to the programboard. Wires light up when the attatched slot is locked.
Fixed a nullpointer exception in Player caused by the testScreen. TestScreen tries to update the sprite, but the sprite doesn't exist.